add explicit type conversion

+++ cpp/models/abm/personal_rng.cpp
@@ -36,7 +36,8 @@ PersonalRandomNumberGenerator::PersonalRandomNumberGenerator(mio::Key<uint64_t> 
 }
 
 PersonalRandomNumberGenerator::PersonalRandomNumberGenerator(const RandomNumberGenerator& model_rng, Person& person)
-    : PersonalRandomNumberGenerator(model_rng.get_key(), person.get_id().get(), person.get_rng_counter())
+    : PersonalRandomNumberGenerator(model_rng.get_key(), static_cast<uint32_t>(person.get_id().get()),
+                                    person.get_rng_counter())
 {
 }
